dto, 도메인 구현 완료
