+++ src/main/java/com/example/jammoney/News/dto/NewsQuizDto.java
@@ -1,4 +1,4 @@
-package com.example.jammoney.dto;
+package com.example.jammoney.News.dto;
 
 import java.util.List;
