+++ src/main/java/com/example/jammoney/News/models/News.java
@@ -1,5 +1,6 @@
-package com.example.jammoney.models;
+package com.example.jammoney.News.models;
 
+import com.example.jammoney.financeTerm.models.FinancialTerm;
 import jakarta.persistence.*;
 import lombok.*;
