+++ src/main/java/com/example/jammoney/News/models/NewsQuiz.java
@@ -1,4 +1,4 @@
-package com.example.jammoney.models;
+package com.example.jammoney.News.models;
 
 import jakarta.persistence.*;
 import lombok.*;
