+++ src/main/java/com/example/jammoney/ThemeLearning/repository/LearningTopicRepository.java
@@ -1,6 +1,6 @@
-package com.example.jammoney.repository;
+package com.example.jammoney.ThemeLearning.repository;
 
-import com.example.jammoney.models.LearningTopic;
+import com.example.jammoney.ThemeLearning.models.LearningTopic;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LearningTopicRepository extends JpaRepository<LearningTopic, Long> {
