+++ src/main/java/com/example/jammoney/financeTerm/dto/UserSavedTermDto.java
@@ -1,4 +1,4 @@
-package com.example.jammoney.dto;
+package com.example.jammoney.financeTerm.dto;
 
 import java.util.List;
