--- src/main/java/com/example/jammoney/service/ThemeService.java
@@ -1,4 +1,0 @@
-package com.example.jammoney.service;
-
-public class ThemeService {
-}
